Make the key module return a helper function & have it handle objects along the key path changing
The key module is now just a function that returns a new Observation.

When it’s given a key path that has multiple objects along it (e.g. inner.key), it will now correctly update the dependency/mutation data if one of the objects changes (e.g. inner gets set to a new object).